refactor(frontend): extract ShellOverlays component from EditorLayout
